# api/openapi.yaml
openapi: 3.0.3
info:
  title: Lalaji — B2B Kirana Shipping Estimator API
  description: |
    REST API for calculating shipping charges in a B2B e-commerce marketplace.
    Kirana store owners can use this to find the nearest warehouse for a seller and
    estimate shipping costs including transport mode selection and delivery speed surcharges.

    ## How shipping charge is calculated
    1. **Find nearest warehouse** — given seller location, Haversine distance to all active warehouses
    2. **Transport mode** — selected by distance band (Mini Van: 0–100 km · Truck: 100–500 km · Aeroplane: 500+ km)
    3. **Billable weight** — `max(actual_weight, L×W×H÷5000)`
    4. **Total charge** — `₹10 base + (rate × distance × billable_weight) [+ ₹1.2/kg for express]`

    ## Error handling
    All errors are returned in the standard response envelope with a machine-readable `errorCode`.
  version: 1.0.0
  contact:
    name: Lalaji Engineering
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://your-render-app.onrender.com
    description: Production (Render)
...
paths:
...
  /api/v1/warehouse/nearest:
    get:
      tags: [Warehouse]
      summary: Get nearest warehouse for a seller
      description: |
        Given a seller ID and product ID, returns the active warehouse that is geographically
        nearest to the seller's dispatch location (Haversine great-circle distance).

        The product must belong to the specified seller.
      operationId: getNearestWarehouse
      parameters:
        - name: sellerId
          in: query
          required: true
          description: Unique ID of the seller (must be a positive integer)
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
        - name: productId
          in: query
          required: true
          description: Unique ID of the product to ship (must belong to the given seller)
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        "200":
          description: Nearest warehouse found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearestWarehouseResponse"
              example:
                success: true
                requestId: "550e8400-e29b-41d4-a716-446655440000"
                data:
                  warehouseId: 1
                  warehouseName: BLR_Warehouse
                  warehouseLocation:
                    lat: 12.99999
                    lng: 77.58264
                  distanceKm: 48.32
        "400":
          description: |
            Missing or invalid parameters.
            - `sellerId` or `productId` is missing
            - Either value is not a positive integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_params:
                  summary: Missing sellerId
                  value:
                    success: false
                    requestId: "550e8400-e29b-41d4-a716-446655440000"
                    error: "sellerId and productId are required query parameters"
                invalid_id:
                  summary: Non-integer sellerId
                  value:
                    success: false
                    requestId: "550e8400-e29b-41d4-a716-446655440000"
                    error: "sellerId must be a positive integer"
        "404":
          description: Seller or product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                seller_not_found:
                  summary: Seller ID does not exist
                  value:
                    success: false
                    requestId: "550e8400-e29b-41d4-a716-446655440000"
                    error: "seller with id 999 not found or is not active"
                product_not_found:
                  summary: Product does not belong to seller
                  value:
                    success: false
                    requestId: "550e8400-e29b-41d4-a716-446655440000"
                    error: "product with id 5 not found for seller 1"
        "503":
          description: No active warehouses available on the platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                requestId: "550e8400-e29b-41d4-a716-446655440000"
                error: "no active warehouses are currently available; please try again later"
...
components:
  schemas:
    LatLng:
      type: object
      properties:
        lat:
          type: number
          format: double
          example: 12.99999
        lng:
          type: number
          format: double
          example: 77.58264
    NearestWarehouseData:
      type: object
      properties:
        warehouseId:
          type: integer
          format: int64
          example: 1
        warehouseName:
          type: string
          example: BLR_Warehouse
        warehouseLocation:
          $ref: "#/components/schemas/LatLng"
        distanceKm:
          type: number
          format: double
          description: Distance (Haversine) from seller to this warehouse in km
          example: 48.32
...
    NearestWarehouseResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          format: uuid
        data:
          $ref: "#/components/schemas/NearestWarehouseData"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        requestId:
          type: string
          format: uuid
        error:
          type: string
          example: "seller with id 999 not found or is not active"
...